+++ src/promptflow/promptflow/azure/_restclient/flow_service_caller.py
@@ -8,7 +8,7 @@ import os
 import sys
 import time
 import uuid
-from functools import wraps
+from functools import wraps, cached_property
 
 import pydash
 
@@ -90,7 +90,7 @@ class FlowServiceCaller(RequestTelemetryMixin):
     DEFAULT_BASE_REGION = 'westus2'
     AML_USE_ARM_TOKEN = 'AML_USE_ARM_TOKEN'
 
-    def __init__(self, workspace, credential, base_url=None, region=None, **kwargs):
+    def __init__(self, workspace, credential, operation_scope, base_url=None, region=None, **kwargs):
         """Initializes DesignerServiceCaller."""
         if 'get_instance' != sys._getframe().f_back.f_code.co_name:
             raise UserErrorException(
@@ -107,7 +107,7 @@ class FlowServiceCaller(RequestTelemetryMixin):
             base_url = os.environ.get(self.FLOW_CLUSTER_ADDRESS, default=base_url)
 
         self._workspace = workspace
-
+        self._operation_scope = operation_scope
         self._service_endpoint = base_url
         self._credential = credential
         retry_policy = RetryPolicy()
@@ -154,6 +154,17 @@ class FlowServiceCaller(RequestTelemetryMixin):
         decoded_token = jwt.decode(token.token, options={"verify_signature": False})
         user_object_id, user_tenant_id = decoded_token["oid"], decoded_token["tid"]
         return user_object_id, user_tenant_id
+
+    @cached_property
+    def _common_azure_url_pattern(self):
+        operation_scope = self._operation_scope
+        pattern = (
+            f"/subscriptions/{operation_scope.subscription_id}"
+            f"/resourceGroups/{operation_scope.resource_group_name}"
+            f"/providers/Microsoft.MachineLearningServices"
+            f"/workspaces/{operation_scope.workspace_name}"
+        )
+        return pattern
 
     @_request_wrapper()
     def create_flow(